main update
all.sh:
-	Now detect other wireless cards that are manager by wpa_supplicant.
 I using a external dongle and when the wireless card is put in p2p mode the interface name is not change it, but we can use "iw dev type" to identify the mode.

D2.py:
- Add overload parameters  :
interface ip address
interface mask
interface name
neighbor peer address
	invoke  ( d2.py -i <self_ip_address> -m <mask> -d <device> -p <ip_peer_address>)
This fix some problems, sometimes the connection tried to establish through another network card, now when the sock is created the interface name is passed to ensure the connection is open to the correct interface

- Timeouts and payloads modified in the socket objects
The sockets has delimiter the timeout and the payload size these changes permit optimize the load and quality in the connection. Also when you press Ctr+^C or the System send a signal to close the proccess to shutdown the program can ear it and exit in correct order.

- Integrated DHCP server.
Now is not need use a external DHCP to assign Ip address to a peer neighbor. The code has integrate "dhcpserver.py" this is a modified version of "PyPXE" proyect.
https://github.com/pypxe/PyPXE
The original code was modified to assignen P2P addres "/30" and asing always the same address to diferents neigbors.
This permit a exact control of address assigned and permit renew the leased time of dhcp servers.

The dhcpserver.py is controller by threading in main program.

-Catch exceptions
Now the code has a complete Catch exceptions and will finish right if a exception is detected. All sockets will be close and the all subprocess will be end

-Debug
level_name='info'
Implemented python logging. Now you can define the information level to debug the code, or put in "info" level to normal use.

-lazycast.service
As now the main program only management one connection per launch.
The lazycast service restart the application when detect a exit code.
Restart=always
RestartSec=2s
lazycast.service was modified to intergrate the daemon  in the main systemctl.

- h264.c
Was suppressed the message "sound error " when only video is transmitter.